Rebase with master and some refactoring for Tags
