+++ datahub-graphql-core/src/main/resources/gms.graphql
@@ -50,6 +50,8 @@ enum EntityType {
     """
     TAG
     """
+    The MLModel Entity
+    """
     MLMODEL
 }
